compose: Add "image" sub-command
The purpose of the "image" sub-command is to allow a user
to create a raw disk-image from an ostree branch to be used
by libvirt. This allows the user to test an ostree branch
by making it bootable.

The way that it works is that the user provides a debos
configuration file in order to create a raw disk. A sample
configuration file is included in this commit. In order
to create a disk image apt-ostree will create a scratch
ostree repository and copy the desired branch into
the scratch repo. Once that happens debos will run
with a specific branch chosen by the user.

Debos is a golang based tool that allows the creation of
Debian based OS images simpler.

To build an image from a provided configuration file, one
simply runs the following command:

   sudo apt-ostree compose image --repo <path to ostree repo> \
     --base <path to configuration directory> \
     <ostree branch>

More details can be found in the man page.

Testing:
PASSED Installed apt-ostree from git repo.
PASSED Run "apt-ostree compose image --base config/debian/image \
        --repo=/repo test"
PASSED Checked for raw disk image in
       /var/tmp/apt-ostree/build/test/image

Story: 2010867
Task: 48556

Depends-On: https://review.opendev.org/c/starlingx/apt-ostree/+/890704

Change-Id: I4ce64214dc3bb59ef03b35c2c27def017ea35487
Signed-off-by: Charles Short <charles.short@windriver.com>