Ensure directories are created in /var
According to the ostree documentation "/var" is a state
directory and it should remain empty by default. To ensure
that the directories are created when the image starts we
use systemd-tmpfiles to create the directories.

The directories are symlinked from /usr/rootdirs/var when
the image is created. The list of directories that are suppose
to be symlinked are generated from that directory.

In order to generate the list, the apt package cache is queried
to check which packages write to the "/var" directory. Packages
to check can be excluded by a list as well.

In the future if new packages are added and the package creates
an entry in /var, the directory will be moved to /usr/rootdirs/var
and the list will be generated again during post-configure.

Also update the gate configuration to install python3-apt from
bullseye and use sitepackages since the apt python module is only
available via Debian package not via pypi.

Depends-On: https://review.opendev.org/c/starlingx/apt-ostree/+/890704

Story: 2010867
Task: 48556

Testing:
PASSED Installed apt-ostree from git repo.
PASSED Run "apt-ostree compose create --base config/debian/bookworm \
       --repo=/tmp/test debian/bookworm"
PASSED Checked out debian/bookworm branch
PASSED Checked for usr/lib//tmpfiles.d/ostree-integration-autovar.conf

Change-Id: I2569978fc948e6352edb03acdf1e4766345b02c4
Signed-off-by: Charles Short <charles.short@windriver.com>