Add "repo" command
Add a "repo" command to create the Debian package archive.
The purpose of the "repo" command is that allows the user
to create, manage, and add new Debian packages locally which
can be used to update ostree branches if the user does
not have access to a network.

The way that the "repo" command works is that, the "repo"
makes system calls to reprepro to manage the archive.
Reprepro was chosen because its simple to use, proven,
and there is plenty of documentation.

In-order for the package feed for it to be effective, one
usually has a web front end so that apt can interact
with the repository. To set one up please read the following
documentaiton:

https://wiki.debian.org/DebianRepository/SetupWithReprepro

This commits also adds an "init" sub-command which will
initialize the configuration of a Debian package repository
created by reprepro.

Testing:
PASSED Installed apt-ostree from git repo.
PASSED Run "sudo apt-ostree repo init \
	--origin starlingx-updates --release bookworm"
PASSED Check for /var/repository directory on target.
PASSED Check "/var/repository/conf/distributions" to see
       the "Origin:" matches up with the CLI.

Story: 2010867
Task: 48556

Change-Id: I9fd56ab78ec6d164b4486b4ef7ad3592dae56ed7
Signed-off-by: Charles Short <charles.short@windriver.com>