repo: "add" subcommand
The purpose of the "add" subcommand is to add additional
Debian packages to a previously created Debian archive
repository.

The way this works is that when the add subcommand
runs, it makes a system call to reprepro. The command
that runs is reprepro with the includedeb option.

To add packages to the repository, you have to download
the packages locally to a directory and then run
"sudo apt-ostree repo add --release <release> *.deb".
Packages can be downloaded via "apt-get download <package>".
Please note, if the package is new and not found in the
regular Debin archive then run-time dependencies needed
to be added as well. Otherwise the package install
will fail when a user tries to install the package via
apt.

Testing:
PASSED Installed apt-ostree from git repo.
PASSED Download desired debian package
PASSED Run "sudo apt-ostree repo add \
      --release bookworm *.deb"
PASSED Check the output of
 "reprepo -b <repo dir> list

Story: 2010867
Task: 48556

Depends-On: https://review.opendev.org/c/starlingx/apt-ostree/+/891591

Change-Id: Ic773899fa57062e60c9e09d05eb383522c33e35c
Signed-off-by: Charles Short <charles.short@windriver.com>