gate: Fix gate
Sphinx 7.2.0 was released today which broke the gate, and prevented
reviews from merging. So use install the previous version of sphinx
in the venv to get things moving again.

Also the flake8 gate seem to have broken as well, this was due
to a change in the tox.ini that prevented zuul from running it.

Story: 2010867
Task: 48556

Change-Id: I6c3531221a105c4dbb8bf6ae310fb5d3c4457f6d
Signed-off-by: Charles Short <charles.short@windriver.com>