compose: Add "compose exec" command
The purpose of the "compose exec" command is to provide
a user to run a one-shot command inide of an ostree
branch.

The way that this works is that a branch is checked
out to a temporary rootfs. Then bubblewrap executes
a the desired command in a seperate namespace. The
namespace will have the "/var", "/usr/etc", and "/usr"
bind mounted from the temporary location to be used
to execute the command sent via the CLI. After the command
has been executed, the temoprary rootfs is checked for
new files and added to a systemd-tmpfiles configuration file
to be executed at the next boot. Finally, the rootfs is
commited back to the ostree repository.

Testing
PASSED Install apt-ostree via pip.
PASSED Create a ostree branch via apt-ostree compose create.
PASSED Install the libvirtd-daemon-system debian package from
       the Debian package archive.

	sudo apt-ostree compose exec --repo /path-to-repo \
		--branch test_branch \
		apt-get install libvirt-daemon-system
PASSED After intsallation check for the /var/log/libvirt
       directory in /var/tmp/apt-ostree/test_branch/<commit>/
       usr/rootdirs/var/log/libvirt

Story: 2010867
Task: 48556

Depends-On: https://review.opendev.org/c/starlingx/apt-ostree/+/892386

Change-Id: I0b0f1ae0460b5a862fde20fa190771e15e8ca41f
Signed-off-by: Charles Short <charles.short@windriver.com>